Clean: Separate funcs from cmd_struct_remove_quotes.c to two files
