+++ minishell.h
@@ -6,7 +6,7 @@
 /*   By: mbartos <mbartos@student.42prague.com>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/12/02 16:24:52 by aldokezer         #+#    #+#             */
-/*   Updated: 2024/03/18 12:54:54 by mbartos          ###   ########.fr       */
+/*   Updated: 2024/03/18 16:53:32 by mbartos          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -142,10 +142,13 @@ void	ft_push_rotate_token(t_cmd *cmd, char *token_text, t_type token_type);
 void	ft_delete_cmd(t_cmd *cmd_table);
 void	ft_move_token(t_cmd *cmd, t_token *ptr_token);
 
-// cmd_struct_remove_quotes.c
+// cmd_struct_remove_quotes_utils.c
 char	**remove_quotes_encaptulates_words(char **arr_of_str);
 char	*get_substr_from_word(char const *str, size_t *i);
+int		count_substrs_in_word(char const *text);
 char	**parse_token_text(char *text);
+
+// cmd_struct_remove_quotes.c
 void	remove_quotes_in_token(t_token *ptr_node);
 void	remove_quotes_in_cmd_tokens(t_cmd *ptr_cmd_tab);
